requirements.txt
streamlit==1.28.0
pandas==2.1.0
matplotlib==3.7.0
seaborn==0.12.2
psycopg2-binary==2.9.7
openpyxl==3.1.2
reportlab==4.0.4
python-dotenv==1.0.0
schedule==1.2.0